+++ examples/errors/main.go
@@ -8,5 +8,5 @@ func main() {
 	logger.Err(mrcore.FactoryErrInternal.Caller(-1).New())
 	logger.Err(mrcore.FactoryErrInternalTypeAssertion.Caller(-1).New("MY-TYPE", "MY-VALUE"))
 
-	logger.Info(mrcore.FactoryErrInternalNoticeDataContainer.Caller(-1).New("MY-DATA").Error())
+	logger.Info(mrcore.FactoryErrStorageQueryDataContainer.Caller(-1).New("MY-DATA").Error())
 }